requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cssselect==1.1.0
fake-useragent==0.1.11
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
outcome==1.2.0
parse==1.19.0
psycopg2==2.9.3
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
requests-html==0.10.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.11
w3lib==1.22.0
webdriver-manager==3.8.3
websockets==10.3
wsproto==1.1.0
zipp==3.8.1
